added a TexturedTriangle class, almost completely implemented, just the intersect method is missing

+++ smallPT_projectACG/TextureManager.hpp
@@ -2,7 +2,7 @@
 #define TEXTUREMANAGER_HPP_
 #include <vector>
 #include <string>
-
+#include <cstdint>
 struct Texture
 {
 	std::string name; // maybe not needed
@@ -29,7 +29,7 @@ public:
 	 * for now just 'BMP' support
 	 */
 	int loadTexture(const std::string& filename);
-	const Texture* getByID (uint32_t id) { return textures.size() <= id ? NULL : &textures[id]; }
+	const Texture* getByID (uint16_t id) { return textures.size() <= id ? NULL : &textures[id]; }
 private:
 	std::vector<Texture> textures;
 	// private constructor to prevent instancing
